services/capacity-forecasting/main.py
"""
Capacity Forecasting Service
Predicts system capacity needs based on historical data and trends.
"""

from fastapi import FastAPI
from pydantic import BaseModel
from typing import Dict, Any, List
import pandas as pd
import numpy as np
from datetime import datetime
import nats
import asyncio

app = FastAPI(
    title="Capacity Forecasting Service",
    description="Predicts system capacity needs based on historical data and trends",
    version="0.4.0",
)


class ForecastRequest(BaseModel):
    metrics: List[Dict[str, Any]]
    forecast_horizon_days: int = 7


class ForecastResponse(BaseModel):
    forecast: Dict[str, float]
    confidence_interval: Dict[str, float]
    recommendations: List[str]


@app.get("/healthz")
async def health_check():
    """Health check endpoint"""
    return {"status": "healthy", "service": "capacity-forecasting", "version": "0.4.0"}


@app.get("/")
async def root():
    """Root endpoint with service information"""
    return {
        "service": "capacity-forecasting",
        "version": "0.4.0",
        "description": "Predicts system capacity needs based on historical data and trends",
        "endpoints": ["/healthz", "/forecast", "/metrics"],
    }


@app.post("/forecast", response_model=ForecastResponse)
async def generate_forecast(request: ForecastRequest):
    """Generate capacity forecast based on historical metrics"""

    # Mock forecast calculation (replace with actual ML model)
    forecast = {
        "cpu_usage": 75.5,
        "memory_usage": 68.2,
        "disk_usage": 45.8,
        "network_bandwidth": 82.3,
    }

    confidence_interval = {
        "cpu_usage": 5.2,
        "memory_usage": 7.1,
        "disk_usage": 3.4,
        "network_bandwidth": 8.7,
    }

    recommendations = [
        "Consider scaling up CPU resources within 3 days",
        "Network bandwidth may need upgrade within forecast horizon",
        "Disk usage is within acceptable limits",
    ]

    return ForecastResponse(
        forecast=forecast,
        confidence_interval=confidence_interval,
        recommendations=recommendations,
    )


@app.get("/metrics")
async def get_metrics():
    """Prometheus-style metrics endpoint"""
    return {
        "forecasts_generated_total": 42,
        "forecast_accuracy_percent": 87.5,
        "last_forecast_timestamp": datetime.utcnow().isoformat(),
    }


if __name__ == "__main__":
    import uvicorn

    uvicorn.run(app, host="0.0.0.0", port=8080)
