services/capacity-forecasting/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
clickhouse-driver==0.2.6
requests==2.31.0
nats-py==2.6.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
statsmodels==0.14.0
python-dateutil==2.8.2